Move commands setup to its own file
